minor bug fixes
