+++ src/main/java/io/github/mzmine/gui/mainwindow/MainMenu.fxml
@@ -329,7 +329,7 @@
 		<MenuItem text="Neutral loss plot" onAction="#runModule"
 			userData="io.github.mzmine.modules.visualization.neutralloss.NeutralLossVisualizerModule" />
 		<MenuItem text="Combined module" onAction="#runModule"
-			userData="io.github.mzmine.modules.visualization.combinedModule.CombinedModule" />
+			userData="io.github.mzmine.modules.visualization.combinedmodule.CombinedModule" />
 		<MenuItem text="m/z scan histogram" onAction="#runModule"
 			userData="io.github.mzmine.modules.visualization.mzhistogram.MZDistributionHistoModule" />
 		<MenuItem text="Diagnostic fragmentation filtering"
