+++ src/main/java/io/github/mzmine/modules/visualization/combinedmodule/AxisType.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2006-2020 The MZmine Development Team
+ *
+ * This file is part of MZmine.
+ *
+ * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
+ * General Public License as published by the Free Software Foundation; either version 2 of the
+ * License, or (at your option) any later version.
+ *
+ * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with MZmine; if not,
+ * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+ * USA
+ */
+
+package io.github.mzmine.modules.visualization.combinedmodule;
+
+public enum AxisType {
+  RETENTIONTIME("Retention time"), //
+  PRECURSORIONMZ("Precursor ion m/z"), //
+  PRODUCTIONMZ("Product ion m/z"), //
+  NEUTRALLOSS("Neutral loss");
+
+  private String type;
+
+  AxisType(String type) {
+    this.type = type;
+  }
+
+  public String toString() {
+    return type;
+  }
+}
