+++ src/main/java/io/github/mzmine/modules/visualization/combinedmodule/CombinedModule.java
@@ -16,7 +16,7 @@
  * USA
  */
 
-package io.github.mzmine.modules.visualization.combinedModule;
+package io.github.mzmine.modules.visualization.combinedmodule;
 
 import io.github.mzmine.datamodel.MZmineProject;
 import io.github.mzmine.modules.MZmineModuleCategory;
@@ -31,6 +31,7 @@ public class CombinedModule implements MZmineRunnableModule {
 
   private static final String MODULE_NAME = "MS/MS data visualizer";
   private static final String MODULE_DESCRIPTION = "Scatter plot for MS/MS data of 3 modules (neutralloss,productionfilter & msms)";
+  private CombinedModuleVisualizerWindowController controller;
 
   @Override
   public @Nonnull
@@ -49,7 +50,8 @@ public class CombinedModule implements MZmineRunnableModule {
   public ExitCode runModule(@Nonnull MZmineProject project, @Nonnull ParameterSet parameters,
       @Nonnull Collection<Task> tasks) {
 
-    CombinedModuleVisualizerWindow newWindow = new CombinedModuleVisualizerWindow(parameters);
+    CombinedModuleVisualizerWindow newWindow = new CombinedModuleVisualizerWindow();
+    controller.setParameters(parameters);
     newWindow.show();
 
     return ExitCode.OK;
