+++ src/main/java/io/github/mzmine/modules/visualization/combinedmodule/CombinedModuleVisualizerWindow.fxml
@@ -2,22 +2,33 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ToolBar?>
+<?import javafx.scene.control.Tooltip?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Pane?>
 <BorderPane fx:id="mainPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity"
   minWidth="-Infinity" prefHeight="570.0" prefWidth="844.0" xmlns="http://javafx.com/javafx/8"
   xmlns:fx="http://javafx.com/fxml/1"
-  fx:controller="io.github.mzmine.modules.visualization.combinedModule.CombinedModuleVisualizerWindowController">
+  fx:controller="io.github.mzmine.modules.visualization.combinedmodule.CombinedModuleVisualizerWindowController">
   <center>
     <Pane fx:id="centerPane" prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER"/>
   </center>
   <right>
     <ToolBar fx:id="toolbar" orientation="VERTICAL" prefHeight="570.0" prefWidth="74.0"
       BorderPane.alignment="CENTER">
-      <items>
-        <Button fx:id="highlightPrecursorBtn" mnemonicParsing="false"/>
-      </items>
+      <Button fx:id="highlightPrecursorBtn" mnemonicParsing="false">
+        <graphic>
+          <ImageView fitHeight="20.0" fitWidth="19.0" pickOnBounds="true" preserveRatio="true">
+            <Image url="@../../../../../../../resources/icons/datapointsicon.png"/>
+          </ImageView>
+        </graphic>
+        <tooltip>
+          <Tooltip text="Highlight precursor m/z range..."/>
+        </tooltip>
+      </Button>
+
     </ToolBar>
   </right>
   <bottom>
